# src/swagger_files/QuickCode.Turuncu.UserManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Turuncu - UserManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/user-manager-module/permission-groups/count": {
      "get": {
        "tags": [
          "PermissionGroups"
        ],
        "operationId": "api-user-manager-module-permission-groups-count-get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/permission-groups/{id}": {
      "get": {
        "tags": [
          "PermissionGroups"
        ],
        "operationId": "api-user-manager-module-permission-groups--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionGroupsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PermissionGroups"
        ],
        "operationId": "api-user-manager-module-permission-groups--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGroupsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGroupsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGroupsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PermissionGroups"
        ],
        "operationId": "api-user-manager-module-permission-groups--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/permission-groups/{permissionGroupsId}/portal-permission-groups/{portalPermissionGroupsId}": {
      "get": {
        "tags": [
          "PermissionGroups"
        ],
        "operationId": "api-user-manager-module-permission-groups--permissionGroupsId--portal-permission-groups--portalPermissionGroupsId--get",
        "parameters": [
          {
            "name": "permissionGroupsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "portalPermissionGroupsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionGroupsPortalPermissionGroups_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/portal-permission-groups/count": {
      "get": {
        "tags": [
          "PortalPermissionGroups"
        ],
        "operationId": "api-user-manager-module-portal-permission-groups-count-get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/portal-permission-groups/{id}": {
      "get": {
        "tags": [
          "PortalPermissionGroups"
        ],
        "operationId": "api-user-manager-module-portal-permission-groups--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalPermissionGroupsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PortalPermissionGroups"
        ],
        "operationId": "api-user-manager-module-portal-permission-groups--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionGroupsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionGroupsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionGroupsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PortalPermissionGroups"
        ],
        "operationId": "api-user-manager-module-portal-permission-groups--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/portal-permission-groups/get-portal-permission-groups/{portalPermissionGroupsPermissionGroupId}": {
      "get": {
        "tags": [
          "PortalPermissionGroups"
        ],
        "operationId": "api-user-manager-module-portal-permission-groups-get-portal-permission-groups--portalPermissionGroupsPermissionGroupId--get",
        "parameters": [
          {
            "name": "portalPermissionGroupsPermissionGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortalPermissionGroupsGetPortalPermissionGroupsResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/portal-permissions/count": {
      "get": {
        "tags": [
          "PortalPermissions"
        ],
        "operationId": "api-user-manager-module-portal-permissions-count-get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/portal-permissions/{id}": {
      "get": {
        "tags": [
          "PortalPermissions"
        ],
        "operationId": "api-user-manager-module-portal-permissions--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalPermissionsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PortalPermissions"
        ],
        "operationId": "api-user-manager-module-portal-permissions--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PortalPermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PortalPermissions"
        ],
        "operationId": "api-user-manager-module-portal-permissions--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/portal-permissions/{portalPermissionsId}/portal-permission-groups/{portalPermissionGroupsId}": {
      "get": {
        "tags": [
          "PortalPermissions"
        ],
        "operationId": "api-user-manager-module-portal-permissions--portalPermissionsId--portal-permission-groups--portalPermissionGroupsId--get",
        "parameters": [
          {
            "name": "portalPermissionsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "portalPermissionGroupsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalPermissionsPortalPermissionGroups_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/portal-permission-types/{portalPermissionTypesId}/portal-permission-groups/{portalPermissionGroupsId}": {
      "get": {
        "tags": [
          "PortalPermissionTypes"
        ],
        "operationId": "api-user-manager-module-portal-permission-types--portalPermissionTypesId--portal-permission-groups--portalPermissionGroupsId--get",
        "parameters": [
          {
            "name": "portalPermissionTypesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "portalPermissionGroupsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalPermissionTypesPortalPermissionGroups_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PermissionGroupsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PermissionGroupsPortalPermissionGroups_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionTypeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PortalPermissionGroupsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionTypeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PortalPermissionGroupsGetPortalPermissionGroupsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionTypeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PortalPermissionsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "itemType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PortalPermissionsPortalPermissionGroups_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "portalPermissionTypeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}